refactor: be - File Interceptor 분리 #198
